api/requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12
Flask-Cors==3.0.2
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
packaging==16.8
PyJWT==1.4.2
pyparsing==2.2.0
six==1.10.0
Werkzeug==0.11.15
